backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
bcrypt
python-jose[cryptography]
python-dotenv

